Разбила код на модули
